+++ squirrel_planning_system/CMakeLists.txt
@@ -28,6 +28,7 @@ catkin_package(
 ## include_directories(include)
 include_directories(
   ${catkin_INCLUDE_DIRS}
+  include
 )
 
 ## val sources
